fix: revoke refresh token method now deletes secret (#1603)
* fix lack of secret deletion

* Changed value from mock to instance

* deps

+++ edc-extensions/dataplane/dataplane-token-refresh/token-refresh-core/src/main/java/org/eclipse/tractusx/edc/dataplane/tokenrefresh/core/DataPlaneTokenRefreshServiceImpl.java
@@ -270,6 +270,11 @@ public class DataPlaneTokenRefreshServiceImpl implements DataPlaneTokenRefreshSe
     }
 
     private Result<Void> deleteTokenData(AccessTokenData tokenData) {
+        var deletionResult = vault.deleteSecret(tokenData.id());
+        if (deletionResult.failed()) {
+            return deletionResult;
+        }
+
         var result = accessTokenDataStore.deleteById(tokenData.id());
         if (result.failed()) {
             return Result.failure(result.getFailureDetail());
